requirements.txt
# Azure AI Assistant - Dependencies
# Install with: pip install -r requirements.txt

openai>=1.30.0          # Official OpenAI SDK (supports Azure OpenAI natively)
python-dotenv>=1.0.0    # Loads environment variables from .env file
fastapi>=0.115.0        # API layer for backend integration
uvicorn>=0.30.0         # ASGI server for FastAPI
